+++ src/autoload.php
@@ -6,6 +6,7 @@ $class_map = array(
 	'TripleA_Payment_Gateway_For_WooCommerce\WPPB\WPPB_Object' => __DIR__ . '/vendor/WPPB/class-wppb-object.php',
 	'TripleA_Payment_Gateway_For_WooCommerce\Admin\Admin' => __DIR__ . '/Admin/class-admin.php',
 	'TripleA_Payment_Gateway_For_WooCommerce\Admin\Plugins_Page' => __DIR__ . '/Admin/class-plugins-page.php',
+	'TripleA_Payment_Gateway_For_WooCommerce\Includes\I18n' => __DIR__ . '/Includes/class-i18n.php',
 	'TripleA_Payment_Gateway_For_WooCommerce\WooCommerce\Payment_Gateways' => __DIR__ . '/WooCommerce/class-payment-gateways.php',
 	'TripleA_Bitcoin_Ecommerce_for_WooCommerce_Payment' => __DIR__ . '/WooCommerce/triplea-payment-gateway-main-class.php',
 );
